runner: add tablequeryjs to reports to allow textual filtering
This commit adds tablequeryjs [1] to stb-tester runner reports. This is
useful because it allows Wireshark-like text-based filtering of the HTML
table, executing queries such as:

- `tag ILIKE slave-1 or tag ILIKE slave-2`
- `"exit status" ILIKE "^(1|2)$"`

Currently tablequeryjs depends on both jQuery and jQuery UI. Removing
the latter as a dependency is tracked in [2] and I hope to get around to
this soon. There are also numerous usability improvements, all tracked
in issues, that I'm hoping to implement soon as well.

[1] https://github.com/asimihsan/tablequeryjs

[2] https://github.com/asimihsan/tablequeryjs/issues/4