runner: Display statistics of a job
Display basic stats of a job: total number of tests, number of passed
tests, number of UITestFailures, number of other failures and success
rate.

Counters are updated on the client side, they only include the
visible rows of the table.

Stats are displayed in index.html above the results table and below the
filter input field.