irnetproxy.py: "--listen-address" is "-l", not "-i"
Similarity with common command-line tools:

netcat uses "-l" to mean "listen" (though netcat's "-l" doesn't take an
argument; there's a separate, non-optional, positional "hostname"
argument).

telnet uses "-s" to specify a source address, but that is something
different that an address that a server binds to.

varnish, an http proxy, uses "-a address[:port]". I assume that "a"
stands for "address"; that's fine for varnish which only needs to
specify one address, but our irnetproxy needs to specify two addresses,
the address to listen on and the address of the real irNetBox. So I've
chosen "-l" as it's more explicitly the *listen* address.