test_irnetbox_proxy: Isolated component test for irnetproxy.py
Using irnetproxy.py in between 2 `stbt run` clients and a fake irnetbox,
the test checks that each client gets the right responses.

Note that tests/fake-irnetbox now returns the sequence number in the
OUTPUT_IR_ASYNC ACK message in little-endian format. This is the
opposite of what the spec† says. This matches the implementation in
irnetproxy.py, which has been tested against 2 separate irNetBox-III
devices.

† http://www.redrat.co.uk/products/IRNetBox_Comms-V3.X.pdf