stbt.*_motion: default parameters for motion testing are set in stbt.conf
Much like the templatematching functions, the parameters for `detect_motion`
and `wait_for_motion` can now be set in the users STBT_CONFIG_FILE (or else
are read from the system installed config file).

This gives users flexibility, particularly for the `noise_threshold`
parameter whose value should depend on the quality of the video source.
For example, an analogue source will contain more incidental noise than
HD video, and therefore a lower `noise_threshold` value should be used.
Being able to set a global parameter which is then applied to every call
to the motion detection algorithm (unless overridden) means that a custom
value does not need to be applied to every in-script call explicitly.

[Committer's note: The motivation is that we have seen several
`wait_for_motion` calls raise `MotionTimeout` when there was in fact
motion. We want to be able to change the default `noise_threshold` and
`consecutive_frames` without breaking backwards compatibility with other
users of stb-tester.]