stbt.*_motion: global parameters for motion testing are set in stbt.conf
Much like the templatematching functions, the parameters for `detect_motion`
and `wait_for_motion` can now be set in the users STBT_CONFIG_FILE (or else
are read from the system installed config file).

This gives users flexibility, particularly for the `noise_threshold`
parameter whose value should depend on the quality of the video source.
For example, an analogue source will contain more incidental noise than
HD video, and therefore a lower `noise_threshold` value should be used.
Being able to set a global parameter which is then applied to every call
to the motion detection algorithm (unless overridden) means that a custom
value does not need to be applied to every in-script call explicitly.

We decided that this method was preferential over implementing a
`MotionParameters` class, analogous to the `MatchParameters` class,
because it would only contain 2 parameters, whereas the purpose of the
`MatchParameters` class was to contain the 5 parameters of the template
matching algorithm. This helped tidy up the function definitions
(particularly user defined helper functions) and decreased the maintenance
burden of having to keep all these extra functions in sync with each other
(in terms of keeping the default values the same).