Makefile: Install extra/runner
It is inconvenient that the full path to 'stb-tester.git/extra/runner' has
to be specified to execute runner, and the readme file has to be generated
manually from restructured text.

The command 'make install-runner' copies runner to a permanent location,
compiles README.rst and makes it available as 'man runner'.

The scripts 'run', 'report' and 'server' are installed to <prefix>/bin,
therefore they will be available as commands, without typing the full path.