stbt.py: Add --restart-source command line option
The configuration option 'global.restart_source' is undocumented in
stb-tester (it's only mentioned in the release notes and a comment in
the system-wide config file) even though this is an important feature
for Hauppauge HD PVR users.

To make users aware of this option, make `restart_source` available as
`--restart-source` command line argument, just like other configuration
options, and mention it in the stbt man page.