stbt.py: new function get_frame_timestamp() added
`stbt.get_frame()` returns the last processed frame, but not its timestamp.
Now that `process_all_frames()` has been added so as to take precise
measurements, there needs to be a way of retrieving this timestamp, as well
as the timestamp from the live stream (already implemented).

This way, a usage exmaple would be:

with stbt.process_all_frames():
    t0 = stbt.get_live_stream_timestamp(retry=True)
    stbt.press('MENU')
    stbt.wait_for_match('menu.png')
    t1 = stbt.get_frame_timestamp()
print 'Time spent in launching the menu:', float(t1 - t2) / (10 ** 9)