Rename irnetproxy.py to irnetbox-proxy
To include the full name of the irNetBox in the program's name, and
removing the `.py` extension for consistency with our other executable
utilities (stbt-run, etc).

The only changes to the file apart from renaming it, are: (use the `-M`
option to `git diff` or `git log` to see just what has changed)

* renaming instances of "irnetproxy.py" inside the file itself (in
  docstrings).
* Adding a magic comment (on line 2) so that Emacs knows that this is
  a python file.

Note that irnetproxy_test.py can't use `import` to import a file without
a `.py` extension, so we have to use `imp.load_source`.

A note on the change to `.gitignore`: Importing irnetbox-proxy (as
opposed to running it from the command line) leaves behind a python
byte-compiled file with the `c` suffix. This normally results in a
`.pyc` extension, but since `irnetbox-proxy` doesn't end in `.py`,
the result is `irnetbox-proxyc`.