stbt.py: process_all_frames must ensure there is incoming video
Due to the fact the `process_all_frames()` creates the Queue with unlimited
size from scratch, makes it possible that a test fails due to NoVideo
just because the first thing it does is a call to `process_all_frames()`.

This commit adds this extra step where the queue must have at least one
element before the yield statement.