stbt.py: reduce memory used when queue is non-leaky
By allocating the queue after the ffmpgecolorspace element, we reduce the
memory consumed when it is non-leaky, that is in a `process_all_frames`
context.

To reduce this even more, I've tried to substitute this queue with a queue2
element. However, what queue2 does is store the buffers in disk AS WELL AS
keep them in memory. Therefore, that doesn't help.

The amount of memory used also depends on the processing speed of the
computer. Currently, on a machine with an Intel Core i5, Fedora 18, and
hardly any process running besides the default services of Fedora, this
consumes an extra ~200MB every second.