stbt.press: Raise exception when irNetBox returns error
I have no idea if our irNetBox is ever replying with NACKs, but if it
did, our scripts would have silently ignored the NACK and proceeded.
With this change, we can be more certain that ignored keypresses are
genuine faults in the system under test, and not failures of the test
infrastructure's infrared emission.

We may need to do a similar fix for the LIRC-based control. The existing
LIRC control implementation will raise an exception for errors at the
socket level, but it doesn't wait for the response from LIRC so it might
ignore application-level errors.

test_that_press_fails_on_irnetbox_error is a new test that passes
correctly with the existing irnetbox.py implementation.

test_that_press_fails_on_irnetbox_nack is a new test that would
fail with the existing irnetbox.py implementation, but this commit
also fixes irnetbox.py to raise an exception in that case.