stbt.press: Wait for irNetBox "Asynchronous IR Output Complete" message
This eliminates one source of latency variability when using a script
like this to measure performance of the system under test:

    press("channel_up")
    start = time.time()
    wait_for_match("new channel.png")
    end = time.time()
    print "duration: %f" % (end - start)

This still has a low precision, limited to how long it takes
`wait_for_match` to process a single frame†. But at least it guarantees
that the time between the system-under-test receiving the infrared
signal and our `start = time()` call, is at most the time it takes for
the confirmation message to travel from the irNetBox to our PC. Since
the irNetBox and the PC are probably on the same local network this
should be minimal (compared to the time it takes a set-top box to change
channels, for example); at least it's more predictable than that same
network delay *plus* the irNetBox's delay in sending the infrared
signal.

This will hold open a TCP connection to the irNetBox until the
asynchronous completion message is received; other clients won't be able
to use the irNetBox during that time because the irNetBox only accepts a
single TCP connection at a time. If this is a problem, use the
`irnetbox-proxy` script in this same directory.

Note that I removed the `sleep(0.5)` from `stbt.IRNetBoxRemote.press`.
This isn't necessary now that we wait for the irNetBox to have sent the
infrared signal before sending the next request. I've tested against my
YouView set-top box, with a series of `stbt.press` calls one after the
other (sending 1 0 3 to change channel, or 2 2 2 to enter the letter "C"
in a text box) and the set-top box registers all keypresses correctly.

† There is work in progress to improve performance measurements, by
processing *every* frame (instead of dropping the frames we can't
process in real time) and by using the frame timestamps from the video
stream instead of wall time. See
https://github.com/drothlis/stb-tester/pull/31