irnetbox.py, irnetbox-proxy: Update comment re. endianness bug
Chris Dodge of RedRat confirmed that the endianness of "sequence number"
field of the "Asynchronous IR Signal Output" response message is a bug.
It can never be fixed, to keep compatibility with existing software.
He will update the documentation.

The sequence number is big-endian in the *request* message, and in the
"Asynchronous IR Output Complete" message (which is a message sent
asynchronously). The endianness bug is in the synchronous response to
the request message.