irnetbox-proxy: Preserve endianness in response message forwarded to client
The "sequence number" field of the following three messages is defined
by the spec† to be big endian:

1. Asynchronous IR Signal Output request (message 48 client->irNetBox)
2. Asynchronous IR Signal Output response (message 48 irNetBox->client)
3. Asynchronous IR Output Complete (message 49 irNetBox->client)

(#2 is sent synchronously, as an ack or nack, as soon as #1 is received.

But there is a bug in the irNetBox device, confirmed by Chris Dodge at
RedRat, such that the sequence number in #2 is actually little endian.
`irnetbox-proxy` used the correct (little endian) format for decoding
the received response, but then re-encoded it incorrectly when sending
the response to the client. The proxy must behave just like a real
irNetBox, so that the same client code can use (or not use) the proxy,
so the proxy must implement the same bugs as the irNetBox.

† http://www.redrat.co.uk/products/IRNetBox_Comms-V3.0.pdf

This bug was found by the test test_irnetbox_proxy, which started
failing after commit 92164a9c.