irnetbox.py: "Asynchronous IR Output Complete" payload contains 4 bytes
Section 6.1.3 of the spec† says "the only data it contains is the
message sequence number of the IR Output message as a ushort" but in
practice the irNetBox sends 4 bytes, not 2, of data in this message.
The last 2 bytes are both "0".

This fixes the error "unpack requires a string argument of length 2"
introduced in 92164a9c. Before 92164a9c we ignored the contents of
this message so this bug didn't manifest.

† http://www.redrat.co.uk/products/IRNetBox_Comms-V3.0.pdf