extra/runner summary report: Upgrade tablequeryjs to v0.1.5
v0.1.4 allows the user to pass callback functions to be called when the
user performs a new search. This will be useful in the next commit,
where we'll display a count of total tests run, passed tests, failures
and errors.

Note that v0.1.4 removes the dependency on jQuery UI. Release notes:
https://github.com/asimihsan/tablequeryjs/releases/tag/v0.1.4

v0.1.5 fixes a bug in v0.1.4.