`make install` & Fedora rpm: Install irnetbox-proxy to $prefix/bin
This tool is essential if you're using the RedRat irNetBox (a
network-controlled infrared emitter) with multiple clients, because the
irNetBox only accepts a single TCP connection at a time. It won't hurt
to install it for users who don't need or care about it. It doesn't have
any dependencies outside of the python standard library.