test_draw_text: Don't continue with test if early steps fail
If the first `stbt-run` fails, the second `stbt-run` is bound to fail
too, so the test would correctly report a failure both with and without
this commit. However the log output from the second `stbt-run` would be
confusing; this way you only see the relevant error from the first
`stbt-run`.