stbt-run: Don't bring `draw_text` into script's namespace
We want to encourage users to explicitly `import stbt` in their scripts,
because it's more pythonic and less "magical", and it allows running
pylint over the scripts. Other recent new functions like `frames` and
`get_config` weren't added to this list for the same reason.

This was added in b6a6c015, but was never part of a tagged release so we
can still remove it from the public API. We'll eventually deprecate
these, and later remove this code altogether, but for now it remains for
compatibility with existing user scripts.