test_draw_text: Use "live" videotestsrc
videotestsrc generates frames as fast as it can, which on a typical
system is much faster than 24 frames per second. The `duration_secs`
parameter of `stbt.draw_frames` acts on the timestamps from the video
stream, not the wall-clock time, so using `duration_secs=3` resulted
in the text being displayed only for a very brief period of time (too
fast to even see it).

Though the previous implementation did work, it might give the
impression that `duration_secs` takes a number in milliseconds.

There are two ways to slow down the videotestsrc to real time: Set the
sink to `sync=true` (we use `sync=false`; see `sink_pipeline` in
tests/stbt.conf), or set videotestsrc's `is-live` property. `sync=true`
causes other problems (due to the delays and jitter introduced by the
image processing) so we use `videotestsrc is-live=true`.