Document `stbt.draw_text` in the official API doc
Note that the change to README.rst is the result of running `make doc`
after the change to api-doc.sh.

I edited the docstring so that it displayed nicely in the generated
manpage (the docstrings and README are in reStructuredText format, so
you need a double linebreak in the source to get a new paragraph in the
output).