extra/runner summary report: Total/pass/fail/error counts
Display basic stats of a job: Total number of tests; number of passed
tests (and pass rate); number of failures (UITestFailures, i.e. the
system-under-test didn't behave as expected); number of errors (i.e.
bugs in the test script or problems with the test infrastructure).

Counters are updated on the client side (in the browser). If you've
filtered the results via the "filter" field, the counters only reflect
the visible rows of the table. The counts are displayed below the filter
input field, as a hint that they are affected by the filter.

Based on a patch by Máté Szendrő.