detect_motion/wait_for_motion: reformat docstrings to use definition list
The parameter names are now subsection headings, along with their default
values. Some line wrapping was needed because of the extra indentation.

Note that the docs for `mask` were altered to match the format of
`consecutive_frames` and `noise_threshold`, although its default value is
not placed into the README dynamically because it does not come from
stbt.conf.