extra/runner: Color points in the HTML report plot by status
This updates the runner's `index.html` report template to color the
points based on the css class of the row in the `#testruns` table. This
is done by having a different jqplot series for each css class and
then specifying a color for each one.

The css class of each row is set by `runner/report.css_class()` and can
be one of three values:
- `success`
- `warning`
- `error`

The colors of the points don't match up with those used in the testruns
table. This is because the latter colors are too light and the resulting
plot seems to be difficult to read.