test_runner_results_server: Fix race condition
It seems that Flask (the python web micro-framework used by
`runner/server`) prints "Running on http://localhost:5787/" before it is
actually ready to accept connections. On Travis CI, a continuous
integration service that runs these tests on a VM in the cloud, even the
`sleep 1` is sometimes not long enough to allow `runner/server` to
start, and the `curl` command fails with "Failed connect to
localhost:5787; Connection refused".

Instead of increasing the sleep, which will increase the test's runtime
every time regardless of how long `runner/server` takes to start, we'll
sleep one second at a time only as long as we need to.

I didn't get rid of `expect_runner_to_say 'Running...` even though it is
no longer strictly necessary, because if the server fails to start for
any other reason, the test's error message will be more relevant.