extra/runner: Summary report: Set iframe height on page load
This line was removed in the previous commit (baeb752c) and added to the
click/keydown handler. But we need to set the iframe's height on page
load too, for Chrome or Safari users who navigate away from the summary
report (e.g. by clicking on one of the log files) and then press the
"back" button. Otherwise after pressing "back" the iframe is at its
default height (only enough to show a few lines of text).

Firefox doesn't need this fix because it handles the "back" button
better (the currently selected row is still blue, the iframe is just as
you had left it).