extra/runner summary report: Fix jerky width change of details pane
Bootstrap's "affix" jQuery plugin started having an annoying behaviour
after baeb752c, where the width of the "details" iframe is different
when the window is at the top-most position, than when you scroll down
slightly.

"affix" is designed for fancy websites where an element moves normally
with the document as you scroll, but then "sticks" at the top once
you've scrolled enough. This is overkill for our needs, and a simple
"position: fixed" does the trick, without the flicker.