stbt-tv: Stream live video through network
Launching 'stbt tv' with '-l' (for "listen") command line argument
streams video through a TCP socket. Used in conjunction with
'stbt control', it makes remote configuration of the set-top-boxes under
test possible.

The choice of TCP protocol is unusual for this application; the reason
behind is that it is easier to configure both the client and the server
compared to using RTP over UDP. As this tool serves for the convenience
of test system administrators, getting low latency is not critical.

TCP port is read from the new 'tv.port' parameter of the stb-tester
config file, or the hard-coded value 4953 is used if it isn't
specified. 4953 is the default value used by 'tcpclientsrc'.

Video bandwidth is read from the 'tv.bandwidth' parameter in bps,
defaults to 1500000 (1.5 Mbps). This value is a rule of thumb
compromise of video quality and network load; users can fine-tune the
bitrate with tools like tcpperf.

The receiving side can watch the video stream with
* vlc:
    `vlc tcp://<host>:<port>`
* gst-launch:
    `gst-launch tcpclientsrc host=<host> port=<port> ! \
        mpegtsdemux ! decodebin2 ! autovideosink`