extra/vm: Cosmetic change to ~vagrant/.config/stbt/stbt.conf
`restart_source` is False by default, so this commit doesn't have a
noticeable effect (unless the user changed `restart_source` in
/etc/stbt/stbt.conf, which is unlikely).

I found myself uncommenting the `#restart_source = False` line when I
wanted to use the HDPVR, but forgetting to change the `False` to `True`.
You shouldn't have to change *two* things (remove the comment *and*
change the value).