tests/run-performance-test.sh: Cleanup to allow testing "stbt run -vv"
This commit doesn't change the behaviour or performance measurements of
the script. It simply allows the script to work even if you manually
edit the script to pass "-vv" to "stbt run".

The pattern "/ Position/" has a space on purpose, so that it only
matches the output of the python script's "print" statement, and not the
debug output from "stbt run" itself, which looks like
"..., position=Position...".