extra/runner summary report: Update chart according to user filter
The "filter" text input allows the user to hide rows from the table;
this commit makes the plot only show the visible rows. This is the same
behaviour as the passed/failed/error totals displayed at the top.

I also moved the plot below the filter input, to imply that the plot
is controlled by the filter.