tests: Remove unnecessary "consecutive_matches"
This was originally set to "consecutive_matches=24" so that each test,
with its match rectangle drawn over the videotestsrc pattern, would
remain visible for 2 seconds to give some reassurance to the human
watching the self-tests run.

Now that the selftests use "fakesink" instead of "ximagesink" (cf2b39dc)
this is no longer necessary, and it slows down the test. By now we have
great confidence in our automated tests so the visual confirmation isn't
necessary.

I left "consecutive_matches=2" in one of the tests (test_wait_for_match)
so that we have at least one place that tests that parameter. (I suppose
a better test would be to use "videotestsrc pattern=ball" and assert
that we don't find a match with consecutive_frames > 1.)