extra/runner report: Do not show "404" while the first test is running
It's quite misleading that there is no index.html in the test results
directory until the first test is completed; users new to `runner`
might think that they entered the wrong address or an error occurred
when they are trying to open the test results.

Generate a row in the table of the index page while a test is being
executed. It has white background with grey text, and the exit status
is "still running".

This is done by skipping `report` if it's run with no arguments or
--classify-only and no exit-status is found in the test run directory,
but executing it if it's run with --html-only.