stbt-tv: force-aspect-ratio in *imagesink
The default behaviour of the sink is to create a window to accommodate
the full size of the video being piped to it, or as large as the display
will permit. However, if the size of the display limits the size of
the window, the video is squashed and the pixel aspect ratio is not
maintained.

By adding "force-aspect-ratio=true", the video is letterboxed [1] inside
the window and the aspect-ratio of the video is maintained. I've checked
that "force-aspect-ratio" is a valid option to all the listed *imagesinks,
however for obvious reasons it is not valid for "fakesink", so it has to
be explicitly added for every "real" sink element.

[1] http://en.wikipedia.org/wiki/Letterboxing_%28filming%29